Move some framework only commons into engine module
