+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/config/ConfigManager.kt
@@ -1,6 +1,6 @@
 package io.kotest.engine.config
 
-import io.kotest.common.mapError
+import io.kotest.engine.mapError
 import io.kotest.core.config.AbstractProjectConfig
 import io.kotest.core.config.ProjectConfiguration
 import io.kotest.mpp.log
