+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/results.kt
@@ -1,17 +1,8 @@
-package io.kotest.common
+package io.kotest.engine
 
 inline fun <A, B> Result<A>.flatMap(fn: (A) -> Result<B>): Result<B> {
    return fold({ fn(it) }, { Result.failure(it) })
 }
-
-@KotestInternal
-fun <A> Result<Result<A>>.flatten(): Result<A> {
-   return when {
-      isSuccess -> getOrThrow()
-      else -> Result.failure(exceptionOrNull()!!)
-   }
-}
-
 
 fun <A> Result<A>.mapError(f: (Throwable) -> Throwable): Result<A> =
    fold({ Result.success(it) }, { Result.failure(f(this.exceptionOrNull()!!)) })
