+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/runBlocking.kt
@@ -1,4 +1,4 @@
-package io.kotest.common
+package io.kotest.engine
 
 expect fun <T> runBlocking(f: suspend () -> T): T
