+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/spec/interceptor/instance/InlineTagSpecInterceptor.kt
@@ -1,6 +1,6 @@
 package io.kotest.engine.spec.interceptor.instance
 
-import io.kotest.common.flatMap
+import io.kotest.engine.flatMap
 import io.kotest.core.config.ProjectConfiguration
 import io.kotest.core.spec.Spec
 import io.kotest.core.test.TestCase
