+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/spec/interceptor/ref/IgnoredSpecInterceptor.kt
@@ -1,6 +1,6 @@
 package io.kotest.engine.spec.interceptor.ref
 
-import io.kotest.common.flatMap
+import io.kotest.engine.flatMap
 import io.kotest.core.annotation.Ignored
 import io.kotest.core.config.ExtensionRegistry
 import io.kotest.core.spec.SpecRef
