+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/test/TestInvocationInterceptor.kt
@@ -6,7 +6,7 @@ import io.kotest.core.test.TestResult
 import io.kotest.core.test.TestScope
 import io.kotest.engine.test.interceptors.TestExecutionInterceptor
 import io.kotest.mpp.Logger
-import io.kotest.mpp.replay
+import io.kotest.engine.concurrency.replay
 import kotlinx.coroutines.coroutineScope
 import kotlin.time.Duration
 import kotlin.time.TimeMark
