+++ kotest-framework/kotest-framework-engine/src/desktopMain/kotlin/io/kotest/engine/runBlocking.kt
@@ -1,4 +1,4 @@
-package io.kotest.common
+package io.kotest.engine
 
 actual fun <T> runBlocking(f: suspend () -> T): T = kotlinx.coroutines.runBlocking { f() }
