+++ kotest-framework/kotest-framework-engine/src/jsHostedMain/kotlin/io/kotest/engine/runBlocking.kt
@@ -1,3 +1,3 @@
-package io.kotest.common
+package io.kotest.engine
 
 actual fun <T> runBlocking(f: suspend () -> T): T = error("runBlocking is not available on JS")
