+++ kotest-framework/kotest-framework-engine/src/jsMain/kotlin/io/kotest/engine/runBlocking.js.kt
@@ -1,4 +1,4 @@
-package io.kotest.common
+package io.kotest.engine
 
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
@@ -7,7 +7,7 @@ import kotlinx.coroutines.promise
 @OptIn(DelicateCoroutinesApi::class)
 actual fun runPromise(f: suspend () -> Unit) {
    GlobalScope.promise { f() }.catch {
-      console.log(it)
+      io.kotest.common.console.log(it)
       throw it
    }
 }
